pcxt: add per-process kernel stacks and far code segment
Add 4 × 1 KB per-process kernel stacks at 0xF000-0xFFFF in the DS=0
segment, replacing the old shared 2 KB static stack.  PROC_MAX=4, so
each process slot gets its own kernel stack for use when ISR/syscall
entry switches SS:SP away from the user stack.

Stage2 now double-loads the kernel binary: once to DS:0x0600 (for
data access via SS=0) and again to far segment 0x1060 (for code
execution via CS=0x1000).  All symbols are linked at DS=0 addresses;
the far copy has the same offsets so near calls resolve correctly.
VFS code is placed in a far segment after the core copy.

The page pool base is now dynamic — stage2 computes it as the first
page-aligned address after the last code segment and passes it via
mod_info.  target_pcxt.c reads it into i16_page_pool_base before
mm_init().

Fix page_id_linear() return type from uintptr_t (16-bit on i16) to
uint32_t, preventing truncation of page addresses above 64 KB.

Verified: builds cleanly, boots on QEMU (PiPAPo banner, VFS mount,
init starts, scheduler runs).

+++ src/kernel/mm/page.c
@@ -65,8 +65,8 @@ static uint32_t i16_detect_ram(void) {
 /* ── Internal helpers ───────────────────────────────────────────────────────
  */
 
-static inline uintptr_t page_id_linear(page_id_t id) {
-  return (uintptr_t)id * PAGE_SIZE;
+static inline uint32_t page_id_linear(page_id_t id) {
+  return (uint32_t)id * PAGE_SIZE;
 }
 
 static inline page_id_t linear_page_id(uintptr_t addr) {
